header column types always string
